[BPK-3747] Support custom items in horizontal nav
In order to allow custom items to be used in Horizontal Navigation the
API has been evolved. The existing style is still preferred and
available in the same way however custom items can now be used.
